pyproject.toml
[tool.poetry]
name = "python-config"
version = "0.1.0"
description = "Spring-like python app configuration utility"
authors = ["Roland Molnar <roland@dextra.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
pyyaml = "^6.0.1"
pytest = "^8.2.2"
python-box = "^7.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
